Move data to config file (make arb immutable)
This move the commands (pipelines logic) that are contained inside the
main script to the file containing the variables which is sourced by
main. As both are to be heavily adapted by users, this leave the main
script 'immutable' and it can be updated to newer versions without
messing with the users configuration.